simple writing dict to json and reading back
