+++ jlog.txt
@@ -1,0 +1,1 @@
+{"100000000000000000000000": {"species": "cat", "first_seen": "2021-12-08 13:49:40.985094", "last_seen": "2021-12-08 13:49:40.985110", "num_reads": "0"}, "100000000000000000000001": {"species": "horse", "first_seen": "2021-12-08 13:49:40.985113", "last_seen": "2021-12-08 13:49:40.985114", "num_reads": "0"}, "100000000000000000000002": {"species": "horse", "first_seen": "2021-12-08 13:49:40.985116", "last_seen": "2021-12-08 13:49:40.985118", "num_reads": "0"}}
